feat(onboarding-flow): create use-onboarding-flow hook
